=tes #17080 Move TestUtils to testkit.SocketUtil
